Slide the 8 second segment window in 1 second increments, as in the original implementation
model:clearState() is required for the LookupTable, was not have been applied correctly for the DataParallelTable (multi-GPU) model. Sampling in between training iterations was not allowing the network to converge.

Fixed batch size. Small batch at the end was harming the network and resulting samples. Possibly due to noisier gradient steps. Re-evaluate later as the noisy samples may have been a result of the lack of model:clearState()

Now generating samples very close to the reference implementation. Looking into how to implement weight normalization for Linear and the GRU layers in Torch.